refactor : 도메인 내 하위 패키지 분리
