+++ src/main/java/domain/card/CardNumberType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import static util.ExceptionConstants.*;
