+++ src/main/java/domain/generator/CardRandomGenerator.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.generator;
+
+import domain.card.Card;
+import domain.card.Deck;
 
 public class CardRandomGenerator implements RandomGenerator<Card> {
     private final Deck deck = new Deck();
