+++ src/main/java/domain/generator/RandomGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.generator;
 
 public interface RandomGenerator<T> {
