+++ src/main/java/domain/participant/Participant.java
@@ -1,4 +1,7 @@
-package domain;
+package domain.participant;
+
+import domain.card.Card;
+import domain.card.Cards;
 
 public abstract class Participant {
     protected final Cards cards;
