+++ src/main/java/domain/participant/Player.java
@@ -1,8 +1,9 @@
-package domain;
+package domain.participant;
 
+import domain.card.Cards;
 import java.util.Objects;
 
-public class Player extends Participant{
+public class Player extends Participant {
     private final String name;
 
     public Player(String name, Cards cards) {
