+++ src/main/java/domain/result/GameResultStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public enum GameResultStatus {
     WIN,
