+++ src/test/java/domain/participant/DealerTest.java
@@ -1,7 +1,11 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.Card;
+import domain.card.CardNumberType;
+import domain.card.CardType;
+import domain.card.Cards;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
