+++ src/test/java/domain/participant/PlayersTest.java
@@ -1,7 +1,13 @@
-package domain;
+package domain.participant;
 
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
+import domain.card.Card;
+import domain.card.CardNumberType;
+import domain.card.CardType;
+import domain.card.Cards;
+import domain.result.GameResult;
+import domain.result.GameResultStatus;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
