+++ src/test/java/domain/result/GameResultStatusTest.java
@@ -1,6 +1,7 @@
-package domain;
+package domain.result;
 
 import static org.assertj.core.api.Assertions.*;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
