+++ src/test/java/domain/result/GameResultTest.java
@@ -1,7 +1,9 @@
-package domain;
+package domain.result;
 
 import static org.assertj.core.api.Assertions.*;
 
+import domain.card.Cards;
+import domain.participant.Player;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
